Allow header instances in message delivery methods
If you had multiple headers with the same name and different values, it was impossible to provide them as a one dimensional hash.

This patch changes `Header::listFromArray()` so that it will accept instances of `Header` rather than assuming the argument is a map. The side effect is that users can supply a list of `Header` instances, thereby enabling multiple headers with the same name.

Additionally, this patch rejects array members with an empty string. Previously, an assertion error would have been thrown in dev, but silently and erroneously passed to the server in prod.